Rename add-controller
